+++ experiments/SheetManagerPandas.py
@@ -1,9 +1,8 @@
-from datetime import datetime
 from os import path
 
 import pandas
 
-from exceptions import SheetManagerException
+from lib.exceptions import SheetManagerException
 
 """
 Attempting to use pandas for faster, more efficient merging.
