+++ merge/SheetManager.py
@@ -3,7 +3,7 @@ import csv
 from copy import deepcopy
 from os import path
 
-from exceptions import SheetManagerException
+from lib.exceptions import SheetManagerException
 from lib.util import create_new_filename
 
 SEP = '-' * 70
