+++ scrape/FpsScraper.py
@@ -3,7 +3,7 @@ import time
 
 from selenium.common.exceptions import NoSuchElementException
 
-from exceptions import ScraperException
+from lib.exceptions import ScraperException
 from scrape.Scraper import Scraper
